requirements.txt
streamlit
python-dotenv
langchain-groq
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
langchain-classic
langchain-huggingface
chromadb
sentence-transformers
pypdf